Sms_Spam_Detection/requirements.txt
streamlit
firebase-admin
python-dotenv
langchain
langchain-core
langgraph
langchain-google-genai
